+++ config/settings.py
@@ -128,5 +128,5 @@ STATICFILES_DIRS = [
 ]
 
 LOGIN_URL = "login"
-LOGIN_REDIRECT_URL = "home"
+LOGIN_REDIRECT_URL = "dashboard"
 LOGOUT_REDIRECT_URL = "login"