+++ main/urls.py
@@ -2,8 +2,10 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.home, name='home'),
+    path('', views.dashboard, name='dashboard'),
     path('contact/', views.contact, name='contact'),
     path('about/', views.about, name='about'),
     path('profile/', views.profile, name='profile'),
+    path('instruments/', views.instrument_list, name='instrument_list'),
+    path('instruments/add/', views.instrument_create, name='instrument_create'),
 ]